+++ src/com/genir/aitweaks/core/Obfuscated.java
@@ -8,21 +8,25 @@ import java.util.List;
  * Collection of placeholders for obfuscated type names.
  */
 public class Obfuscated {
-    public static class ThreatEvalAI {
+    public enum ShipCommand {
+        TURN_LEFT, TURN_RIGHT,
     }
 
     public interface AutofireManager {
         void advance_AutofireManager(float dt, ThreatEvalAI threatEvalAI, Vector2f missileDangerDir);
     }
 
+    public interface CombatEntity {
+    }
+
     public interface Maneuver {
+        CombatEntity getTarget_Maneuver();
+    }
+
+    public static class ThreatEvalAI {
     }
 
     public static class ApproachManeuver {
-    }
-
-    public enum ShipCommand {
-        TURN_LEFT, TURN_RIGHT,
     }
 
     public static class ShipCommandWrapper {
@@ -41,6 +45,9 @@ public class Obfuscated {
 
         public Maneuver getCurrentManeuver() {
             return null;
+        }
+
+        public void cancelCurrentManeuver() {
         }
     }
 }