+++ src/com/genir/aitweaks/core/utils/extensions/ShipAPI.kt
@@ -6,11 +6,10 @@ import com.fs.starfarer.api.combat.CombatAssignmentType.*
 import com.fs.starfarer.api.impl.campaign.ids.HullMods
 import com.fs.starfarer.combat.ai.BasicShipAI
 import com.fs.starfarer.combat.entities.Ship
+import com.genir.aitweaks.core.Obfuscated
 import com.genir.aitweaks.core.features.shipai.CustomShipAI
-import com.genir.aitweaks.core.state.combatState
 import com.genir.aitweaks.core.utils.times
 import org.lazywizard.lazylib.MathUtils
-import org.lazywizard.lazylib.ext.getFacing
 import org.lazywizard.lazylib.ext.minus
 import org.lwjgl.util.vector.Vector2f
 
@@ -54,8 +53,15 @@ val ShipAPI.attackTarget: ShipAPI?
         // For manually controlled ship, return R-selected target.
         isUnderManualControl -> shipTarget
 
-        // Fall back to using vanilla AI maneuver target.
-        else -> combatState.maneuverTargetTracker[this]
+        basicShipAI != null -> {
+            val ai = basicShipAI!! as Obfuscated.BasicShipAI
+            val maneuver: Obfuscated.Maneuver? = ai.currentManeuver
+
+            maneuver?.target_Maneuver as? ShipAPI
+        }
+
+        // Fall back to using vanilla maneuver target flag.
+        else -> aiFlags?.getCustom(ShipwideAIFlags.AIFlags.MANEUVER_TARGET) as? ShipAPI
     }
 
 val ShipAPI.isAutomated: Boolean
@@ -66,7 +72,7 @@ val ShipAPI.deploymentPoints: Float
 
 /** Angle between ship facing and direction from ship to point p. */
 fun ShipAPI.angleFromFacing(p: Vector2f): Float {
-    return MathUtils.getShortestRotation((p - location).getFacing(), facing)
+    return MathUtils.getShortestRotation((p - location).facing, facing)
 }
 
 /** Calculates the effective ship velocity in the global frame