+++ src/com/genir/aitweaks/core/utils/extensions/Vector2f.kt
@@ -1,8 +1,11 @@
 package com.genir.aitweaks.core.utils.extensions
 
+import com.genir.aitweaks.core.utils.RADIANS_TO_DEGREES
 import com.genir.aitweaks.core.utils.Rotation
+import com.genir.aitweaks.core.utils.atan2
+import org.lazywizard.lazylib.MathUtils
 import org.lazywizard.lazylib.VectorUtils
-import org.lazywizard.lazylib.ext.getFacing
+import org.lazywizard.lazylib.ext.isZeroVector
 import org.lazywizard.lazylib.ext.minus
 import org.lazywizard.lazylib.ext.plus
 import org.lwjgl.util.vector.Vector2f
@@ -29,7 +32,11 @@ fun Vector2f.addLength(toAdd: Float): Vector2f {
 }
 
 val Vector2f.facing: Float
-    get() = getFacing()
+    get() {
+        if (this.isZeroVector()) return 0f
+
+        return MathUtils.clampAngle(atan2(this.y, this.x) * RADIANS_TO_DEGREES)
+    }
 
 val Vector2f.length: Float
     get() = length()